commit working hashmap & generic value linkedlist
Was a bit bored so started looking at how hashmaps are actually formed. This is an attempt at a fixed size hashmap with separate chaining to work around hash collisions. The linked list implementation has the string key being used when inserting in to the hashmap so help identify the specific <T> value being stored where a collision on the string key has occured and it is instead inserted in to the linked list, but we can jump closer to where the value is stored via the key being hashed and that result being the index for the hashmap buckets. The value stored is of interface{} type so that we can store more complicated values with a key, such as instances of structs. It is then down the the implementer to ensure type safety or to know exactly what is in each map (for now).